rename modules and fix tests
